companion/requirements.txt
pysher
requests
python-dotenv
